fix: resolve max lines eslint
