+++ src/constants/variables.ts
@@ -1,2 +1,3 @@
 export const BIG_SORT_NUMBER: number = 999;
 export const LIMIT_PREVENT_INFINITE_LOOPS: number = 200;
+export const REGEX_GROUP_STRING: string = `\\(['"\`](.*?)['"\`]\\)`;
