+++ src/helpers/handlers/getContentSendTestCase.ts
@@ -1,0 +1,20 @@
+import { contentRequestType } from '../../interfaces/extractData';
+import { findValueInCode } from '../findValueInCode';
+
+const GROUP_VALUE_POSITION_SEND_CONTENT: number = 1;
+export const getContentSendTestCase = ({
+  testCaseText,
+  textFileTest,
+  directoryAllTests,
+}: {
+  testCaseText: string;
+  textFileTest: string;
+  directoryAllTests: string;
+}): contentRequestType => {
+  const RE_CONTENT_SEND: RegExp = /\.send\(([^))]{1,9999})[\S\s]{0,500}\)[\S\s]{0,500}[;\\.]/;
+  const contentSend: RegExpExecArray | null = RE_CONTENT_SEND.exec(testCaseText);
+  if (contentSend) {
+    return findValueInCode(contentSend[GROUP_VALUE_POSITION_SEND_CONTENT], textFileTest, directoryAllTests);
+  }
+  return '';
+};
